media: Change usage datatype to uint64 for C2OMXNode
 - The usage datatype is uint64 at CCodec and there is possibility
   of loosing bits which are set to > 31 due to trauncation and
   datatype of usage variable is uint32 at C2OMXNode.
 - Updated require changes to avoid this truncation.

+++ media/libstagefright/omx/1.0/WGraphicBufferSource.cpp
@@ -143,7 +143,7 @@ Return<Status> TWGraphicBufferSource::configure(
 
     // use consumer usage bits queried from encoder, but always add
     // HW_VIDEO_ENCODER for backward compatibility.
-    uint32_t  consumerUsage;
+    uint64_t  consumerUsage;
     void *_params = &consumerUsage;
     uint8_t *params = static_cast<uint8_t*>(_params);
     fnStatus = UNKNOWN_ERROR;
